+++ src/main/java/Nanali/domain/cody/cloth/Outfit.java
@@ -26,12 +26,12 @@ public class Outfit extends BaseEntity { // Outfit은 코디 사진
     @Enumerated(EnumType.STRING)
     private Sex sex;
 
-    private Long temp_from;
-    private Long temp_to;
-    private Long uv_from;
-    private Long uv_to;
-    private Long rain_from;
-    private Long rain_to;
+    private Double temp_from;
+    private Double temp_to;
+    private Double uv_from;
+    private Double uv_to;
+    private Double rain_from;
+    private Double rain_to;
 
     @OneToMany(mappedBy = "outfit")
     private List<LikeOutfit> outfitList;
@@ -40,8 +40,8 @@ public class Outfit extends BaseEntity { // Outfit은 코디 사진
 
     }
 
-    public Outfit(String imgName, String imgUrl, Style style, Sex sex, Long temp_from, Long temp_to,
-                  Long uv_from, Long uv_to, Long rain_from, Long rain_to) {
+    public Outfit(String imgName, String imgUrl, Style style, Sex sex, Double temp_from, Double temp_to,
+                  Double uv_from, Double uv_to, Double rain_from, Double rain_to) {
         this.imgName = imgName;
         this.imgUrl = imgUrl;
         this.style = style;
