+++ src/main/java/Nanali/repository/GarmentRepository.java
@@ -15,31 +15,31 @@ public interface GarmentRepository extends JpaRepository<Garment, Long> {
             "and g.sex = :sex or g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'OUTER' " +
             "and g.sex = 'BOTH'")
-    List<Garment> findAllByCategoryIsOuter(@Param(value = "temp") Long temp, @Param(value = "uv") Long uv,
-                                           @Param(value = "rain") Long rain, @Param(value = "sex") Sex sex);
+    List<Garment> findAllByCategoryIsOuter(@Param(value = "temp") Double temp, @Param(value = "uv") Double uv,
+                                           @Param(value = "rain") Double rain, @Param(value = "sex") Sex sex);
 
     @Query("select g from Garment g where g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'TOP' " +
             "and g.sex = :sex or g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'TOP' " +
             "and g.sex = 'BOTH'")
-    List<Garment> findAllByCategoryIsTop(@Param(value = "temp") Long temp, @Param(value = "uv") Long uv,
-                                         @Param(value = "rain") Long rain, @Param(value = "sex") Sex sex);
+    List<Garment> findAllByCategoryIsTop(@Param(value = "temp") Double temp, @Param(value = "uv") Double uv,
+                                         @Param(value = "rain") Double rain, @Param(value = "sex") Sex sex);
 
     @Query("select g from Garment g where g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'PANTS' " +
             "and g.sex = :sex or g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'PANTS' " +
             "and g.sex = 'BOTH'")
-    List<Garment> findAllByCategoryIsPants(@Param(value = "temp") Long temp, @Param(value = "uv") Long uv,
-                                           @Param(value = "rain") Long rain, @Param(value = "sex") Sex sex);
+    List<Garment> findAllByCategoryIsPants(@Param(value = "temp") Double temp, @Param(value = "uv") Double uv,
+                                           @Param(value = "rain") Double rain, @Param(value = "sex") Sex sex);
 
     @Query("select g from Garment g where g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'SHOES' " +
             "and g.sex = :sex or g.temp_from <= :temp and g.temp_to >= :temp and g.uv_from <= :uv " +
             "and g.uv_to >= :uv and g.rain_from <= :rain and g.rain_to >= : rain and g.category = 'SHOES' " +
             "and g.sex = 'BOTH'")
-    List<Garment> findAllByCategoryIsShoes(@Param(value = "temp") Long temp, @Param(value = "uv") Long uv,
-                                           @Param(value = "rain") Long rain, @Param(value = "sex") Sex sex);
+    List<Garment> findAllByCategoryIsShoes(@Param(value = "temp") Double temp, @Param(value = "uv") Double uv,
+                                           @Param(value = "rain") Double rain, @Param(value = "sex") Sex sex);
 
 }
